[refactor] : RootLayout 구조 재정비 및 GNB, 배경, Suspense 분리 통합
- framer-motion 기반 BackgroundWrapper 구조 정비 및 경로 기반 스타일 적용
- ClientGNBWrapper를 통한 GNB 조건부 렌더링 적용 (zustand 상태 및 경로 분기)
- SEO metadata 분리(constants)
- next/font/local 기반 pretendard variable 설정
- Suspense로 fallback 로딩 처리 구성 완료
