+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import localFont from "next/font/local";
 import "./globals.css";
 import { defaultMetadata as metadata } from "@/constants/metadata";
 import { Suspense } from "react";
-import ClientGNBWrapper from "@/features/common/components/navbar/clientGnbWrapper";
+import ClientGNBWrapper from "@/features/common/components/clientGnbWrapper";
 import Loading from "./loading";
 import BackgroundWrapper from "@/features/common/ui/BackgroundWrapper";
